requirements.txt
beautifulsoup4==4.9.3
feedparser==6.0.2
html5lib==1.1
pyowm==3.2.0
python-dotenv==0.17.0
python-telegram-bot==13.4.1
pytz==2020.1
requests==2.24.0
